add to collection button on details card, save new collection refresh page
