+++ homepare/src/UserCollections.jsx
@@ -7,6 +7,7 @@ import { DetailsCard } from "./detailsCard";
 import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import axios from "axios";
+import { useNavigate } from "react-router-dom";
 
 export function UserCollections( {token}) {
   const thumbWidth = "100px";
@@ -136,6 +137,7 @@ export function UserCollections( {token}) {
 
 export function NewCollection( {token}) {
   const [collectionInput, setCollectionInput] = useState("");
+  const navigate = useNavigate();
 
   const handleSaveCollection = (e) => {
     console.log("save button clicked");
@@ -151,7 +153,7 @@ export function NewCollection( {token}) {
           authorization: `x-access-token ${token}`
         },
       }
-    );
+    ).then(navigate("/UserCollections"))
   };
 
   return (
